# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is pixel_clash.",
        "title": "pixel_clash",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
        "/player/join": {
            "post": {
                "description": "player joins game",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "player"
                ],
                "summary": "player joins game",
                "parameters": [
                    {
                        "description": "user name, desired capacity",
                        "name": "name",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/types.PostPlayerJoinHandlerRequest"
                        }
                    }
                ],
                "responses": {}
            }
        },
        "/player/status": {
            "post": {
                "description": "sends status info",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "player"
                ],
                "summary": "sends status info",
                "parameters": [
                    {
                        "description": "user id",
                        "name": "name",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/types.PostPlayerStatusHandlerRequest"
                        }
                    }
                ],
                "responses": {}
            }
        }
    },
    "definitions": {
        "types.PostPlayerJoinHandlerRequest": {
            "type": "object",
            "properties": {
                "capacity": {
                    "type": "integer"
                },
                "nickname": {
                    "type": "string"
                }
            }
        },
        "types.PostPlayerStatusHandlerRequest": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                }
            }
        }
    }
}